Subject: Initial Commit
Body:
Adding the main pages to the basic site and creating a working server.
It handles the url params and serves the correct page based on the url.
If the params are incorrect then the page is then directed to the 404.html
page. This is also backed up with another error handler in case this also
fails.